merge next_release

+++ tests/testthat/test-viewAntares.R
@@ -36,4 +36,10 @@ opts <- setSimulationPath(studyPath)
 if(requireNamespace("rhdf5")){
   rhdf5::H5close()
 }
-unlink(tpDir, recursive = TRUE)
+
+if(requireNamespace("rhdf5")){
+  if(dir.exists(tpDir))
+  {
+    unlink(tpDir, recursive = TRUE)
+  }
+}
